requirements.txt
# FastAPI and Uvicorn
fastapi==0.118.0
uvicorn==0.37.0

# Image Upload Dependencies
python-multipart==0.0.9
Pillow==10.4.0

# --- 추가된 라이브러리 ---
# For calling external APIs (CLOVA Speech)
httpx==0.27.0

# For managing environment variables
python-dotenv==1.0.1

# Pydantic is already a dependency of FastAPI, but explicitly listing it is fine
pydantic==2.11.9

# --- gRPC 라이브러리 추가 ---
grpcio==1.62.2
grpcio-tools==1.62.2

# --- 기존 라이브러리 ---
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
colorama==0.4.6
h11==0.16.0
idna==3.10
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0